wrapping state for edited semesters
